requirements.txt
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.0
plotly>=5.15.0
tqdm>=4.65.0
kaggle>=1.5.0
torch>=2.0.0
